backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
alembic==1.13.3
asyncpg==0.29.0
aiosqlite==0.20.0
cryptography==43.0.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0
celery==5.4.0
redis==5.0.8
GitPython==3.1.43
python-dotenv==1.0.1
bcrypt==4.0.1
faker==30.1.0
anyio==4.6.0
pyyaml==6.0.2
jinja2==3.1.4
fpdf2==2.8.1
tree-sitter-java
tree-sitter-cpp
tree-sitter-go
tree-sitter-rust
tree-sitter-ruby
tree-sitter-php
tree-sitter-c-sharp
# Enterprise additions
pgvector>=0.3.6
google-generativeai>=0.8.0
openai>=1.40.0
neo4j>=5.20.0
